Use setTimeout instead of socket.setTimeout
We noticed unexpected linear increases in memcached command latency that
greatly exceeded the configured timeout value.

Digging in, we realized that the issue is that scoket.setTimeout only
triggers the timeoutHandler if the socket has been idle for longer than
the timeout. In our case, commands were piling up and we were running a
huge queue because an issue increased the latency of each command but
not enough to trigger the timeout. All of the commands were long past
their deadlines when they were processed, but the socket was never idle
long enough to trigger the deadline checking.

This switches the command timeout to regular setTimeout since what the
client cares about and wants to guard against is waiting longer than
timeout for the command to complete. The fact that bits are flowing is
irrelevant. This will cause commands to timeout properly and cause the
queue to be flushed, reducing load and allowing new work to be
prioritized over already dead work.